pyproject.toml
[tool.poetry]
name = "openai_handson"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.23.1"
deepl = "^1.11.0"
python-dotenv = "^0.21.0"
isort = "^5.10.1"
black = "^22.10.0"
flake8 = "^5.0.4"
mypy = "^0.982"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
